Added better precision. Changed from double to BigDecimal

+++ src/com/tholix/service/ReceiptManager.java
@@ -1,8 +1,9 @@
 /**
- * 
+ *
  */
 package com.tholix.service;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
@@ -15,13 +16,13 @@ import com.tholix.domain.ReceiptEntity;
 /**
  * @author hitender
  * @when Dec 26, 2012 3:09:48 PM
- * 
+ *
  */
 public interface ReceiptManager extends RepositoryManager<ReceiptEntity> {
 	public static String TABLE = BaseEntity.getClassAnnotationValue(ReceiptEntity.class, Document.class, "collection");
 
 	public List<ReceiptEntity> getAllObjectsForUser(String userProfileId);
-	
+
 	//public List<ReceiptGrouped> getAllObjectsGroupedByDate(String userProfileId);
-	public Map<Date, Double> getAllObjectsGroupedByDate(String userProfileId);
+	public Map<Date, BigDecimal> getAllObjectsGroupedByDate(String userProfileId);
 }
